requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
pydantic==2.6.4
pytest==7.4.0
pytest-mock==3.14.0
httpx==0.27.0
openai==1.10.0
apscheduler
